extraction de consolidateBatch() dans son propre fichier => correction du test associé
